requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-generativeai==0.7.2
python-dotenv==1.0.1
pillow==10.4.0
slowapi==0.1.9
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
